+++ components/date-picker/TimeList.js
@@ -27,14 +27,14 @@ export default class TimeList extends Component {
     const dVal = 32
     this.listRef.current && (this.listRef.current.scrollTop = value * dVal)
   }
-  componentDidMount () {
-    setTimeout(() => {
-      this.addListener()
-      this.scrollTo()
-    }, 0)
-  }
   componentDidUpdate () {
     this.scrollTo()
+  }
+  componentDidMount () {
+    this.listRef.current.addEventListener('scroll', this.scrollEvent)
+    setTimeout(() => {
+      this.scrollTo()
+    }, 0)
   }
   componentWillUnmount () {
     window.clearTimeout(this.timer)
@@ -126,7 +126,7 @@ export default class TimeList extends Component {
         }
         {this.liSuffix}
       </ul>
-      {showArrow && this.renderArrow('hours')}
+      {showArrow && this.renderArrow(type)}
     </div>
   }
 }