+++ components/date-picker/dateUtil.js
@@ -11,7 +11,7 @@ import parse from 'date-fns/parse'
 import toDate from 'date-fns/toDate'
 import startOfWeek from 'date-fns/startOfWeek'
 import endOfWeek from 'date-fns/endOfWeek'
-import dateFormat from 'date-fns/format'
+import format from 'date-fns/format'
 import addMonths from 'date-fns/addMonths'
 import isSameMonth from 'date-fns/isSameMonth'
 import getYear from 'date-fns/getYear'
@@ -46,7 +46,20 @@ const getStartDate = (dateObj) => {
 const getEndDate = (dateObj) => {
   return getValidDate(dateObj.endDate)
 }
-
+const compatibleToDate = (value, format) => {
+  if (typeof value === 'string') {
+    return parse(value, format, new Date())
+  }
+  return toDate(value)
+}
+const compatibleFormatString = (formatStr) => {
+  return formatStr.replace(/[Y+|D+]/g, (word) => {
+    return word.toLowerCase()
+  })
+}
+const dateFormat = (value, formatStr) => {
+  return isValid(value) ? format(value, compatibleFormatString(formatStr)) : ''
+}
 export {
   getDaysInMonth, // 获取当月的天数
   subMonths, // 月份减法
@@ -79,5 +92,7 @@ export {
   getEndDate, // 封装 - 获取结果时间
   getValidDate, // 获取有效的时间
   toDate,
-  parseISO
+  parseISO,
+  compatibleToDate,
+  compatibleFormatString
 }