+++ components/upload/upload-legacy/UploadClick.js
@@ -21,7 +21,7 @@ class UploadClick extends Upload {
       fileCountLimted
     } = this.state
     return (
-      <div className='hi-upload hi-upload--normal'>
+      <div className='hi-upload-legacy hi-upload-legacy--normal'>
         <label>
           <input
             ref={node => { this.uploadRef = node }}
@@ -32,28 +32,28 @@ class UploadClick extends Upload {
             accept={accept}
             hidden
           />
-          <span className={`hi-upload__button ${(disabled || fileCountLimted) ? 'hi-upload__button--disabled' : ''}`}>
+          <span className={`hi-upload-legacy__button ${(disabled || fileCountLimted) ? 'hi-upload-legacy__button--disabled' : ''}`}>
             { buttonText || localeDatas.upload.buttonText}
           </span>
         </label>
         {
-          tips && <span className='hi-upload__tips hi-upload__tips--single-line'>{tips}</span>
+          tips && <span className='hi-upload-legacy__tips hi-upload-legacy__tips--single-line'>{tips}</span>
         }
         {showUploadList && (
-          <ul className='hi-upload__list'>
+          <ul className='hi-upload-legacy__list'>
             {fileList.map((file, index) => {
               const fileNameCls = classNames(
-                'hi-upload__filename',
-                file.uploadState === 'error' && 'hi-upload__filename--error'
+                'hi-upload-legacy__filename',
+                file.uploadState === 'error' && 'hi-upload-legacy__filename--error'
               )
               return (
                 <li
                   key={index}
-                  className='hi-upload__item'
+                  className='hi-upload-legacy__item'
                   title={file.name}
                 >
                   <span className={classNames(`Ficon-${file.fileType}`)} />
-                  <div className='hi-upload__right-content'>
+                  <div className='hi-upload-legacy__right-content'>
                     <span className={fileNameCls} title={file.name}>{file.name}</span>
                     <span>
                       { onRemove &&
@@ -66,8 +66,8 @@ class UploadClick extends Upload {
                   </div>
                   {
                     file.uploadState === 'loading' && (
-                      <div className='hi-upload__upstatus'>
-                        <i className='hi-upload__upstatus-line' style={{ width: file.progressNumber + '%' }} />
+                      <div className='hi-upload-legacy__upstatus'>
+                        <i className='hi-upload-legacy__upstatus-line' style={{ width: file.progressNumber + '%' }} />
                       </div>
                     )
                   }