+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # 更新日志
 
+## 2.5.1
+
+- 修复 `<Input />` defaultValue 设置无效的问题 [#717](https://github.com/XiaoMi/hiui/issues/717)
+- 修复 `<Upload />` 拖拽上传场景下，删除文件不正确的问题 [#705](https://github.com/XiaoMi/hiui/issues/705)
+- 修复 `<Menu />` showAllSubMenus 对传入 data 的容错问题 [#687](https://github.com/XiaoMi/hiui/issues/687)
+- 修复 `<DatePicker />` format 后格式不符合预期的问题 [#712](https://github.com/XiaoMi/hiui/issues/712)
+- 修复 `<TimePicker />` 清空后再次打开面板清空无效的问题 [#722](https://github.com/XiaoMi/hiui/issues/722)
+- 修复 `<Popover />` 在某些场景下显示、隐藏的顺序不正确的问题 [#706](https://github.com/XiaoMi/hiui/issues/706)
+- 修复 `<Tree />` 国际化支持不完全的问题 [#729](https://github.com/XiaoMi/hiui/issues/729)
+- 修复 `<Dropdown />` 数据项 href 仅点击文字才能跳转的问题 [#726](https://github.com/XiaoMi/hiui/issues/726)
+
+## 2.5.0
+
+- 新增 `<Cascader />` filterOption 属性支持自定义搜索 [#704](https://github.com/XiaoMi/hiui/issues/704)
+- 修复 `<Cascader />` 刷新后 value 不能正确显示的问题 [#667](https://github.com/XiaoMi/hiui/issues/667)
+- 修复 `<Pagination />` pageSizeOptions 写法兼容性问题 [#703](https://github.com/XiaoMi/hiui/issues/703)
+- 修复 `<Carousel />` 第一张图向前翻页跳转不正确的问题 [#696](https://github.com/XiaoMi/hiui/issues/696)
+- 修复 兼容属性 legacy 造成的组件污染问题 [#708](https://github.com/XiaoMi/hiui/issues/708)
+- 修复 `<DatePicker />` 传入字符串值时控制台抛出的警告问题 [#709](https://github.com/XiaoMi/hiui/issues/709)
+
+## 2.4.1
+
+- 修复 `<Table />` 多选失效的问题 [#699](https://github.com/XiaoMi/hiui/issues/699)
+
+## 2.4.0
+
+- 修复 `<Select />` 没有适配主题色的问题 [#257](https://github.com/XiaoMi/hiui/issues/257)
+- 修复 `<Tabs />` 空状态下报错的问题 [#674](https://github.com/XiaoMi/hiui/issues/674)
+- 修复 `<Radio />` 非受控状态下切换失效的问题 [#683](https://github.com/XiaoMi/hiui/issues/683)
+- 修复 `<Cascader />` 清空搜索展示不正确的问题 [#333](https://github.com/XiaoMi/hiui/issues/333)
+- 修复 `<Collpase />` disabled 没有样式区分的问题 [#483](https://github.com/XiaoMi/hiui/issues/483)
+- 修复 `<Checkbox />` 非受控情况下 onChange 参数不全的问题 [#688](https://github.com/XiaoMi/hiui/issues/688)
+- 修复 `<Table />` 多选的显示问题 [#685](https://github.com/XiaoMi/hiui/issues/685)
+- 修复 `<Pagination />` 每页大小切换无效的问题 [#691](https://github.com/XiaoMi/hiui/issues/691)
+- 新增 `<Upload />` 增加 loading 状态 [#680](https://github.com/XiaoMi/hiui/issues/680)
+- 新增 `<Upload />` 增加照片墙预览提示 icon [#490](https://github.com/XiaoMi/hiui/issues/490)
+- 新增 `<Select />` filterOption 属性支持自定义搜索 [#673](https://github.com/XiaoMi/hiui/issues/673)
+
 ## 2.3.0
 
 - 修复 `<Checkbox />` 同时渲染多个 context provider 的问题 [#649](https://github.com/XiaoMi/hiui/issues/649)
@@ -40,7 +78,6 @@
 - 修复 `<Input />` clearable 不生效的问题 [#580](https://github.com/XiaoMi/hiui/issues/580)
 - 修复 `<Upload />` 中的自定义上传结合 maxCount 不能上传的问题 [#582](https://github.com/XiaoMi/hiui/issues/582)
 - 修复 `<Input />` type 为 textarea 时的底部的样式问题 [#584](https://github.com/XiaoMi/hiui/issues/584)
-
 
 ## 2.1.0